% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayRmbt_forward.R
\name{bayrmbt_forward}
\alias{bayrmbt_forward}
\title{BAYMBT forward model}
\usage{
bayrmbt_forward(t, Tmodel, Type)
}
\arguments{
\item{t}{A scalar or vector of temperature values (1 x N) or (N x 1)}

\item{Tmodel}{A string corresponding the model you want to use. Options are
"T" = assumes mean annual air temperature (BayMBT); "T0" = assumes mean
annual temperature of months above zero (BayMBT0).}

\item{Type}{A string corresponding to the data type. Options are: "soil" or
"lake". Function will use the respective calibration.
Example: bayRmbt_forward(t,Tmodel="T0",Type="lake")}
}
\value{
A 1000-member ensemble of MBT'5Me values (N x 1000)
}
\description{
Applies the BAYMBT forward model for MBT5Me measured in soils and peats.
Predicts MBT5Me based on mean annual air temperature or mean temperatures
above zero.
}
\references{
Dearing Crampton-Flood, E., Tierney, J. E., Peterse, F., Kirkels, F. M.,
& Sinninghe Damst?, J. S. (2020). BayMBT: A Bayesian calibration model
for branched glycerol dialkyl glycerol tetraethers in soils and peats.
Geochimica et Cosmochimica Acta, 268, 142-159.

Martinez-Sosa, P., Tierney, J. E., Stefanescu, I. C., Dearing
Crampton-Flood, E., Shuman, B. N., Routson, C. (2021) A global Bayesian
temperature calibration for lacustrine brGDGTs.
Geochimica et Cosmochimica Acta, 305, 87-105.
}
