% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayRmbt_model.R
\name{bayrmbt_model}
\alias{bayrmbt_model}
\title{BAYMBT calibration model}
\usage{
bayrmbt_model(Tmodel, Type)
}
\arguments{
\item{Tmodel}{A string corresponding the temperature model you want to
calculate. Options are: "T" or "T0", they calculate the mean annual air
temperature (BayMBT) or the mean annual temperatures above zero (BayMBT0),
respectively.}

\item{Type}{A string corresponding to the data type. Options are "soil" or
"lake". The function uses the respective calibration.}
}
\value{
Generates parameter files for the input type of sample and model. Files are saved as RDS files.
Example: bayRmbt_model(Tmodel="T",Type="lake")
}
\description{
Generates the BAYMBT calibration model for MBT5Me measured in soils and peats.
Computes slope, intercept, error variance terms using Bayesian linear
regression.
}
\references{
Dearing Crampton-Flood, E., Tierney, J. E., Peterse, F., Kirkels, F. M.,
& Sinninghe Damsté, J. S. (2020). BayMBT: A Bayesian calibration model
for branched glycerol dialkyl glycerol tetraethers in soils and peats.
Geochimica et Cosmochimica Acta, 268, 142-159.
}
