% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_FAs_by_set.R
\name{calculate_FAs_by_set}
\alias{calculate_FAs_by_set}
\title{Structural Set Fractional Abundance (FA) Calculation Function}
\usage{
calculate_FAs_by_set(raw, set_type = "all", data = "raw")
}
\arguments{
\item{raw}{A data frame of HPLC-MS peak area values (Nx15+1). Column 1 must
have unique identifiers for each sample. GDGT names should be formatted as per
the publication.}

\item{set_type}{Indicates the set to be calculated. Options are:
"Meth" (methylation set), "Cyc" (cyclization set), "Isom" (Isomer set),
"MI" (Meth-Isom set), "CI" (Cyc-Isom set), "MC" (Meth-Cyc set),
and "Full" (Full set).}

\item{data}{Indicates the type of data being processed. The only accepted
data types are peak areas or concentrations ("raw").}
}
\value{
result: A data frame with additional columns where the indicated
FAs for the selected sets are calculated for each brGDGT for each sample.
}
\description{
Calculate fractional abundances for the structural sets defined in
Raberg et al. (2021): (Meth, Cyc, Isom, MI, CI, MC, Full)
}
\note{
Function structure developed with the aid of Claude AI
}
\references{
Raberg, J. H., Harning, D. J., Crump, S. E., de Wet, G., Blumm, A., Kopf, S., ... & Sepúlveda, J. (2021). Revised fractional abundances and warm-season temperatures substantially improve brGDGT calibrations in lake sediments. Biogeosciences, 18(12), 3579-3603.
}
\author{
Jonathan H. Raberg
}
