% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_single_set.R
\name{calculate_single_set}
\alias{calculate_single_set}
\title{Calculate FAs by sets}
\usage{
calculate_single_set(data, set_name, group_vars)
}
\arguments{
\item{data}{A data frame of HPLC-MS peak area values (Nx15+1). Column 1 must
have unique identifiers for each sample. GDGT names should be formatted as per
the publication.}

\item{set_name}{Indicates the set to be calculated. Options are:
"Meth" (methylation set), "Cyc" (cyclization set), "Isom" (Isomer set),
"MI" (Meth-Isom set), "CI" (Cyc-Isom set), "MC" (Meth-Cyc set),
and "Full" (Full set).}

\item{group_vars}{Indicates the structural variables that will be used to
select the brGDGTs for each set. Options are "methyl", "cycle", and "pos",
following the formatting structure of the format_GDGT_sets function.}
}
\value{
result: A data frame with an additional column where the indicated
FA for a set is calculated for each brGDGT for each sample.
}
\description{
Calculates the FAs of brGDGTs acording to the sets as defined by
Raberg et al., (2022).Data must be formatted with structural information
of the GDGTs using the format_GDGT_sets function.
}
\note{
Function structure developed with the aid of Claude AI
}
\references{
Raberg, J. H., Harning, D. J., Crump, S. E., de Wet, G., Blumm, A., Kopf, S., ... & Sepúlveda, J. (2021). Revised fractional abundances and warm-season temperatures substantially improve brGDGT calibrations in lake sediments. Biogeosciences, 18(12), 3579-3603.
}
\author{
Jonathan H. Raberg
}
