% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBTp.R
\name{CBTp}
\alias{CBTp}
\title{CBT' index}
\usage{
CBTp(gdgt, complete = TRUE, na.ignore = FALSE)
}
\arguments{
\item{gdgt}{Data frame with the peak areas or fractional abundances of brGDGTs. Each row should correspond to a sample and each column to each brGDGT with the appropriated name format.}

\item{complete}{Logic variable. Indicates whether all brGDGTs in the original formulation are present in the data frame (TRUE, default value) or if some are missing. If TRUE the function will not proceed if any variable is missing, it will produce an error. If FALSE, the function will remove the missing variables from the equation.}

\item{na.ignore}{Logic variable. If FALSE (default) the function will turn any NA values into 0 for the calculations.}
}
\value{
The function returns the CBT' value for each sample provided. Any value that blows up (<-10000 or >10000) is turned into NA.
}
\description{
Calculates the CBT' index (De Jonge et al., 2014) for brGDGTs presented as either peak areas or fractional abundances.
}
\references{
De Jonge, C., Hopmans, E. C., Zell, C. I., Kim, J. H., Schouten, S., & Damsté, J. S. S. (2014). Occurrence and abundance of 6-methyl branched glycerol dialkyl glycerol tetraethers in soils: Implications for palaeoclimate reconstruction. Geochimica et Cosmochimica Acta, 141, 97-112.
}
