% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComIdx.R
\name{ComIdx}
\alias{ComIdx}
\title{Community index}
\usage{
ComIdx(gdgt, complete = TRUE, na.ignore = FALSE)
}
\arguments{
\item{gdgt}{Data frame with the peak areas or fractional abundances of brGDGTs. Each row should correspond to a sample and each column to each brGDGT with the appropriated name format.}

\item{complete}{Logic variable. Indicates whether all brGDGTs in the original formulation are present in the data frame (TRUE, default value) or if some are missing. If TRUE the function will not proceed if any variable is missing, it will produce an error. If FALSE, the function will remove the missing variables from the equation.}

\item{na.ignore}{Logic variable. If FALSE (default) the function will turn any NA values into 0 for the calculations.}
}
\value{
The function returns the CI value for each sample provided.
}
\description{
Calculates the Community Index (De Jonge et al., 2019) for brGDGTs presented as either peak areas or fractional abundances.
}
\references{
De Jonge, C., Radujković, D., Sigurdsson, B. D., Weedon, J. T., Janssens, I., & Peterse, F. (2019). Lipid biomarker temperature proxy responds to abrupt shift in the bacterial community composition in geothermally heated soils. Organic Geochemistry, 137, 103897.
}
