% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComoUnGDGT-package.R
\docType{package}
\name{ComoUnGDGT-package}
\alias{ComoUnGDGT}
\alias{ComoUnGDGT-package}
\title{ComoUnGDGT: Suite of functions useful for importing, manipulating, and predicting environmental parameters from GDGTs}
\description{
What the package does (one paragraph).
}
\keyword{internal}
