% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctGs.R
\name{correctGs}
\alias{correctGs}
\title{Generates equations for GDGT indices}
\usage{
correctGs(eq)
}
\arguments{
\item{eq}{Equation for the index that wants to be calculated. This is provided by the index function as a text string.}
}
\value{
The function returns the equation for the index.
}
\description{
Compares the provided list of GDGTs to the required list of GDGTs for the index that one wants to calculate. Any GDGTs not provided in the dataset are removed from the equation and a warning is generated.
}
