% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DegCycp.R
\name{DegCycp}
\alias{DegCycp}
\title{Degrees of Cyclization' index}
\usage{
DegCycp(gdgt, complete = TRUE, na.ignore = FALSE)
}
\arguments{
\item{gdgt}{Data frame with the peak areas or fractional abundances of brGDGTs. Each row should correspond to a sample and each column to each brGDGT with the appropriated name format.}

\item{complete}{Logic variable. Indicates whether all brGDGTs in the original formulation are present in the data frame (TRUE, default value) or if some are missing. If TRUE the function will not proceed if any variable is missing, it will produce an error. If FALSE, the function will remove the missing variables from the equation.}

\item{na.ignore}{Logic variable. If FALSE (default) the function will turn any NA values into 0 for the calculations.}
}
\value{
The function returns the DC' value for each sample provided. Any value that blows up (<-10000 or >10000) is turned into NA.
}
\description{
Calculates the DC' index (De Jonge, 2021) for brGDGTs presented as either peak areas or fractional abundances.
}
\references{
De Jonge, C., Kuramae, E. E., Radujković, D., Weedon, J. T., Janssens, I. A., & Peterse, F. (2021). The influence of soil chemistry on branched tetraether lipids in mid-and high latitude soils: Implications for brGDGT-based paleothermometry. Geochimica et Cosmochimica Acta, 310, 95-112.
}
