% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_GDGT_sets.R
\name{format_GDGT_sets}
\alias{format_GDGT_sets}
\title{Format GDGT data to calculate FAs by sets}
\usage{
format_GDGT_sets(raw)
}
\arguments{
\item{raw}{A matrix of HPLC-MS peak area values (Nx15+1). Column 1 must have
unique identifiers for each sample. GDGT names should be formatted as per
the publication.}
}
\value{
output$prior_mean: User choice of prior mean

raw_data: A data frame with N*15 rows, and 5 additional columns
to the original dataset. Columns are: Compound (brGDGT name),
Amount (concentration or peak area), Methyl (I to III, depending on the
number of methyl groups), Cycle (a to c depending on the number of cyclo
pentanes), and Pos (5 or 6 depending on the position of the methyl groups)
}
\description{
Formats a dataset formatted as a table where columns are GDGT concentrations
or peak areas and rows samples, with a first column containing unique ids,
and transforms the GDGT columns to long format with structural information.
}
\note{
Function structure developed with the aid of Claude AI

Here we're assigning Ia, Ib, and Ic to be "5-methyl" compounds here
in accordance with Raberg et al. (2021)
}
\references{
Raberg, J. H., Harning, D. J., Crump, S. E., de Wet, G., Blumm, A., Kopf, S., ... & Sepúlveda, J. (2021). Revised fractional abundances and warm-season temperatures substantially improve brGDGT calibrations in lake sediments. Biogeosciences, 18(12), 3579-3603.
}
\author{
Jonathan H. Raberg
}
