% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2G3.R
\name{G2G3}
\alias{G2G3}
\title{Ratio of GDGT2 to GDGT3}
\usage{
G2G3(gdgt, complete = TRUE, na.ignore = FALSE)
}
\arguments{
\item{gdgt}{Data frame with the peak areas or fractional abundances of isoGDGTs. Each row should correspond to a sample and each column to each brGDGT with the appropriated name format.}

\item{complete}{Logic variable. Indicates whether all isoGDGTs in the original formulation are present in the data frame (TRUE, default value) or if some are missing. If TRUE the function will not proceed if any variable is missing, it will produce an error. If FALSE, the function will remove the missing variables from the equation.}

\item{na.ignore}{Logic variable. If FALSE (default) the function will turn any NA values into 0 for the calculations.}
}
\value{
The function returns the GDGT2/GDGT3 value for each sample provided.
}
\description{
Calculates the ratio of GDGT2 to GDGT3 fractional abundances of isoGDGTs. Each row should correspond to a sample and each column to each isoGDGT with the appropriated name format.
}
