% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoRat.R
\name{IsoRat}
\alias{IsoRat}
\title{Isomer 6Me Ratio}
\usage{
IsoRat(gdgt, complete = TRUE, na.ignore = FALSE)
}
\arguments{
\item{gdgt}{Data frame with the peak areas or fractional abundances of brGDGTs. Each row should correspond to a sample and each column to each brGDGT with the appropriated name format.}

\item{complete}{Logic variable. Indicates whether all brGDGTs in the original formulation are present in the data frame (TRUE, default value) or if some are missing. If TRUE the function will not proceed if any variable is missing, it will produce an error. If FALSE, the function will remove the missing variables from the equation.}

\item{na.ignore}{Logic variable. If FALSE (default) the function will turn any NA values into 0 for the calculations.}
}
\value{
The function returns the IR6Me value for each sample provided.
}
\description{
Calculates the IR6Me ratio (De Jonge et al., 2015) for brGDGTs presented as either peak areas or fractional abundances.
}
\references{
De Jonge, C., Stadnitskaia, A., Fedotov, A., & Damsté, J. S. S. (2015). Impact of riverine suspended particulate matter on the branched glycerol dialkyl glycerol tetraether composition of lakes: The outflow of the Selenga River in Lake Baikal (Russia). Organic Geochemistry, 83, 241-252.
}
