% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoRatsevn.R
\name{IsoRatsevn}
\alias{IsoRatsevn}
\title{Isomer 7Me Ratio}
\usage{
IsoRatsevn(gdgt, complete = TRUE, na.ignore = FALSE)
}
\arguments{
\item{gdgt}{Data frame with the peak areas or fractional abundances of brGDGTs, extended version. Each row should correspond to a sample and each column to each brGDGT with the appropriated name format.}

\item{complete}{Logic variable. Indicates whether all brGDGTs in the original formulation are present in the data frame (TRUE, default value) or if some are missing. If TRUE the function will not proceed if any variable is missing, it will produce an error. If FALSE, the function will remove the missing variables from the equation.}

\item{na.ignore}{Logic variable. If FALSE (default) the function will turn any NA values into 0 for the calculations.}
}
\value{
The function returns the IR7Me value for each sample provided.
}
\description{
Calculates the IR6Me ratio (Wang et al., 2021) for brGDGTs presented as either peak areas or fractional abundances.
}
\references{
Wang, H., Liu, W., He, Y., Zhou, A., Zhao, H., Liu, H., ... & Liu, Z. (2021). Salinity-controlled isomerization of lacustrine brGDGTs impacts the associated MBT5ME'terrestrial temperature index. Geochimica et Cosmochimica Acta, 305, 33-48.
}
