% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearCalib.R
\name{linearCalib}
\alias{linearCalib}
\title{Application of calibrations for GDGTs}
\usage{
linearCalib(
  gdgt,
  depth = NA,
  calibration,
  data = "raw",
  complete = TRUE,
  na.ignore = FALSE
)
}
\arguments{
\item{gdgt}{Data frame with the peak areas or fractional abundances of GDGTs. Each row should correspond to a sample and each column to each GDGT with the appropriated name format.}

\item{depth}{Water depth in meters for the Stefanescu temperature calibration (Eq. 10). 1 by default. Depth must be a vector with equal length as the number of samples provided.}

\item{calibration}{Selection of the forward model to be used. Current options are presented in Suplementary Table 1 of the (Terrestrial GDGT review).}

\item{data}{Data is presented as peak areas ("raw") or fractional abundances ("frac"). Default is peak areas.}

\item{complete}{Logic variable. Indicates whether all brGDGTs in the original formulation are present in the data frame (TRUE, default value) or if some are missing. If TRUE the function will not proceed if any variable is missing, it will produce an error. If FALSE, the function will remove the missing variables from the equation.}

\item{na.ignore}{Logic variable. If FALSE (default) the function will turn any NA values into 0 for the calculations.}
}
\value{
The function returns the MBT'5Me value for each sample provided.
}
\description{
Applies published forward models based on linear functions (Eq. 1 to 15) for reconstructing temperature using GDGTs
}
