% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetIdx.R
\name{MetIdx}
\alias{MetIdx}
\title{Methane Index}
\usage{
MetIdx(gdgt, complete = TRUE, na.ignore = FALSE)
}
\arguments{
\item{gdgt}{Data frame with the peak areas or fractional abundances of isoGDGTs. Each row should correspond to a sample and each column to each isoGDGT with the appropriated name format.}

\item{complete}{Logic variable. Indicates whether all isoGDGTs in the original formulation are present in the data frame (TRUE, default value) or if some are missing. If TRUE the function will not proceed if any variable is missing, it will produce an error. If FALSE, the function will remove the missing variables from the equation.}

\item{na.ignore}{Logic variable. If FALSE (default) the function will turn any NA values into 0 for the calculations.}
}
\value{
The function returns the MI value for each sample provided.
}
\description{
Calculates the Methane index (Zhang, 2011) for isoGDGTs presented as either peak areas or fractional abundances.
}
\references{
Zhang, Y. G., Zhang, C. L., Liu, X. L., Li, L., Hinrichs, K. U., & Noakes, J. E. (2011). Methane Index: A tetraether archaeal lipid biomarker indicator for detecting the instability of marine gas hydrates. Earth and Planetary Science Letters, 307(3-4), 525-534.
}
