% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rib.R
\name{Rib}
\alias{Rib}
\title{Ri/b index}
\usage{
Rib(gdgt, complete = TRUE, na.ignore = FALSE)
}
\arguments{
\item{gdgt}{Data frame with the peak areas or fractional abundances of GDGTs. Each row should correspond to a sample and each column to each GDGT with the appropriated name format.}

\item{complete}{Logic variable. Indicates whether all GDGTs in the original formulation are present in the data frame (TRUE, default value) or if some are missing. If TRUE the function will not proceed if any variable is missing, it will produce an error. If FALSE, the function will remove the missing variables from the equation.}

\item{na.ignore}{Logic variable. If FALSE (default) the function will turn any NA values into 0 for the calculations.}
}
\value{
The function returns the Ri/b value for each sample provided.
}
\description{
Calculates Ri/b (Xie et al., 2012) for GDGTs presented as either peak areas or fractional abundances.
}
\references{
Xie, S., Pancost, R. D., Chen, L., Evershed, R. P., Yang, H., Zhang, K., ... & Xu, Y. (2012). Microbial lipid records of highly alkaline deposits and enhanced aridity associated with significant uplift of the Tibetan Plateau in the Late Miocene. Geology, 40(4), 291-294.
}
