% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3a2a.R
\name{S3a2a}
\alias{S3a2a}
\title{Ratio of IIIa to IIa}
\usage{
S3a2a(gdgt, complete = TRUE, na.ignore = FALSE)
}
\arguments{
\item{gdgt}{Data frame with the peak areas or fractional abundances of brGDGTs. Each row should correspond to a sample and each column to each brGDGT with the appropriated name format.}

\item{complete}{Logic variable. Indicates whether all brGDGTs in the original formulation are present in the data frame (TRUE, default value) or if some are missing. If TRUE the function will not proceed if any variable is missing, it will produce an error. If FALSE, the function will remove the missing variables from the equation.}

\item{na.ignore}{Logic variable. If FALSE (default) the function will turn any NA values into 0 for the calculations.}
}
\value{
The function returns the SIIIa/SIIa value for each sample provided.
}
\description{
Calculates the ratio proposed by Xiao et al., (2016) of IIIa to IIa for brGDGTs presented as either peak areas or fractional abundances.
}
\references{
Xiao, W., Wang, Y., Zhou, S., Hu, L., Yang, H., & Xu, Y. (2016). Ubiquitous production of branched glycerol dialkyl glycerol tetraethers (brGDGTs) in global marine environments: a new source indicator for brGDGTs. Biogeosciences, 13(20), 5883-5894.reference
}
