% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEX86.R
\name{TEX86}
\alias{TEX86}
\title{TEX86 index}
\usage{
TEX86(gdgt, complete = TRUE, na.ignore = FALSE)
}
\arguments{
\item{gdgt}{Data frame with the peak areas or fractional abundances of isoGDGTs. Each row should correspond to a sample and each column to each brGDGT with the appropriated name format.}

\item{complete}{Logic variable. Indicates whether all 6 isoGDGTs are present in the data frame (TRUE, default value) or if some are missing. If TRUE the function will not proceed if any variable is missing, it will produce an error. If FALSE, the function will remove the missing variables from the equation.}

\item{na.ignore}{Logic variable. If FALSE (default) the function will turn any NA values into 0 for the calculations.}
}
\value{
The function returns the TEX86 value for each sample provided.
}
\description{
Calculates the TEX86 index (Schouten, 2002) for isoGDGTs presented as either peak areas or fractional abundances.
}
\references{
Schouten, S., Hopmans, E. C., Schefuß, E., & Damste, J. S. S. (2002). Distributional variations in marine crenarchaeotal membrane lipids: a new tool for reconstructing ancient sea water temperatures?. Earth and Planetary Science Letters, 204(1-2), 265-274.
}
