% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayRmbt_predict.R
\name{bayrmbt_predict}
\alias{bayrmbt_predict}
\title{BAYMBT prediction model}
\usage{
bayrmbt_predict(
  gdgt,
  prior_mean,
  prior_std,
  Tmodel,
  Type,
  data = "raw",
  complete = TRUE,
  na.ignore = FALSE
)
}
\arguments{
\item{gdgt}{A matrix of HPLC-MS peak area values (15xN) or (Nx15), or a
scalar vector of MBT'5Me values (1xN) or (Nx1).}

\item{prior_mean}{A scalar prior mean value of T in degrees C.}

\item{prior_std}{A scalar prior standard deviation value of T in degrees C.}

\item{Tmodel}{A string corresponding the temperature model you want to
calculate. Options are: "T" or "T0", they calculate the mean annual air
temperature (BayMBT) or the mean annual temperatures above zero (BayMBT0),
respectively.}

\item{Type}{A string corresponding to the data type. Options are "soil" or
"lake". The function uses the respective calibration.}

\item{data}{A string corresponding the type of data to run in the model.
Options are "raw" or "mbt5", which use peak areas or fractional abundances,
or MBT'5Me values, respectively.}

\item{complete}{Logic variable. Indicates whether all 15 brGDGTs are present in the data frame (TRUE, default value) or if some are missing. If TRUE the function will not proceed if any variable is missing, it will produce an error. If FALSE, the function will remove the missing variables from the equation.}

\item{na.ignore}{Logic variable. If FALSE (default) the function will turn any NA values into 0 for the calculations.
Example: bayRmbt_predict(mbt5me,10,10,Tmodel="T0",Type="lake",data="area",complete=TRUE)}
}
\value{
output$prior_mean: User choice of prior mean

output$prior_std: User choice of prior standard deviation

output$T: 2.5%, 50%, and 97.5% confidence intervals of posterior SST

output$ens: full ensemble of posterior SST (N x 1000);
}
\description{
Predicts Mean annual air temperature or mean temperatures above zero from MBT5Me measured in soils and peats.
}
\references{
Dearing Crampton-Flood, E., Tierney, J. E., Peterse, F., Kirkels, F. M.,
& Sinninghe Damsté, J. S. (2020). BayMBT: A Bayesian calibration model
for branched glycerol dialkyl glycerol tetraethers in soils and peats.
Geochimica et Cosmochimica Acta, 268, 142-159.

Martinez-Sosa, P., Tierney, J. E., Stefanescu, I. C., Dearing
Crampton-Flood, E., Shuman, B. N., Routson, C. (2021) A global Bayesian
temperature calibration for lacustrine brGDGTs.
Geochimica et Cosmochimica Acta, 305, 87-105.
}
