% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayspR_tex.R
\name{bayspR_tex}
\alias{bayspR_tex}
\title{BaySPR TEX86}
\usage{
bayspR_tex(
  gdgt,
  lon,
  lat,
  prior_std,
  runname,
  varargin = 1000,
  data = "raw",
  complete = TRUE,
  na.ignore = FALSE
)
}
\arguments{
\item{gdgt}{A matrix of HPLC-MS peak area values (6xN) or (Nx6), or a
scalar vector of TEX86 values (1xN) or (Nx1).}

\item{lon}{Longitude from -180 to 180}

\item{lat}{Latitude from -90 to 90}

\item{prior_std}{Prior standard deviation}

\item{runname}{Specify which model to use. Enter either "SST" or "subT"}

\item{varargin}{Used to set the number of posterior draws of parameters to
mix across, and whether to save the 5th/50th/95th percentiles of the
predictions, or the whole ensemble as well. If left empty, then 1000 draws
are used and the ensemble of predictions is not saved. If only one argument,
it is the number of draws (cannot exceed 15000), and the ensemble is not
saved. Note that samples are thinned to use the full span of the ensemble.
If two arguments, the first gives the number of draws, while the second is
an indicator: 0 (default) or 1, will save the 5th/50th/95th percentiles
only, or the full ensemble, respectively. In all the above, the total
number of ensemble members is capped by the number of draws in the model
output.}

\item{data}{A string corresponding the type of data to run in the model.
Options are "raw" or "tex", which use peak areas or fractional abundances,
or TEX86 values, respectively.}

\item{complete}{Determines how is TEX86 calculated. TRUE calculates exactly
as per Schouten, et al., 2002 (default). FALSE omits missing compounds. A
warning is generated and the omitted compounds are shown.
prior mean is set as the mean over all instrumental SST observations
within max_dist (distance is chordal), or the closest min_num points:
whichever has the larger number of observations.
To use the closest K points, set min_num=K and max_dist=0.
to use all points within L km, set min_num=1 (ensure at least one) and
max_dist=L. Results from the paper are with min_num=1; max_dist=500;}
}
\value{
Predictions - Nd by 3 array, where Nd=length(dats), giving the 5/50/95
percentiles of the predictions.

SiteLoc   - Location of dats, as the inputted [lon, lat];

GridLoc 	- Location of the grid centroid closest to dats, used to pull the alpha and beta
samples.

PriorMean - The prior mean as calculated above.

PriorStd  - The prior std as input.

PredsEns  - Nd by Nsamps array of predictions. Only include if
ens_sel==1.
}
\description{
Predict T from TEX86 using the standard model. The assumption in this case is that oceanographic conditions are sufficiently similar to today so that it is reasonable to use the spatial distribution of regression parameters as fitted from the core tops.
}
\references{
Tierney, J. E., & Tingley, M. P. (2014). A Bayesian, spatially-varying calibration model for the TEX86 proxy. Geochimica et Cosmochimica Acta, 127, 83-106.

Schouten, S., Hopmans, E. C., Schefuß, E., & Damste, J. S. S. (2002). Distributional variations in marine crenarchaeotal membrane lipids: a new tool for reconstructing ancient sea water temperatures?. Earth and Planetary Science Letters, 204(1-2), 265-274.
}
