% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FracA.R
\name{FracA}
\alias{FracA}
\title{Calculate Fractional Abundance}
\usage{
FracA(gdgt, group, coerce = FALSE, how = "each", na.ignore = FALSE)
}
\arguments{
\item{gdgt}{Data frame with the peak areas of GDGTs. Each row should correspond to a sample and each column to each GDGT with the appropriated name format.}

\item{group}{GDGTs to be transformed into fractional abundances. Options are "iso" (isoGDGTs), "br" (brGDGTs), "br_extended" (5/6 isomer and 7-methyl), "isoGMGTs", or "OHGDGTs".}

\item{coerce}{Logic variable. If FALSE (default) the function will use the provided data. If TRUE, the function will incorporate missing GDGTs from each group as 0s for the calculations.}

\item{how}{Select how the fractional abundance will be calculated. Options are "each" (default), which calculates the fractional abundance of each group independently; and "all", which calculates the fractional abundance of each GDGT over the total sum of selected GDGTs.}

\item{na.ignore}{Logic variable. If FALSE (default) the function will turn any NA values into 0 for the calculations.}
}
\value{
The function returns the fractional abundances of all GDGTs selected for all samples provided.
}
\description{
Calculates the fractional abundance of the selected GDGTs presented as peak areas.
}
