% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearCalib.R
\name{linearCalib}
\alias{linearCalib}
\title{Application of linear calibrations for GDGTs}
\usage{
linearCalib(
  gdgt,
  env,
  calibration,
  data = "raw",
  complete = TRUE,
  na.ignore = FALSE
)
}
\arguments{
\item{gdgt}{Data frame with the peak areas or fractional abundances of GDGTs. Each row should correspond to a sample and each column to each GDGT with the appropriated name format.}

\item{env}{Selection of the environmental parameter to reconstruct. Options currently are "Temperature" and "pH"}

\item{calibration}{Selection of the forward model to be used. Currently options are "Russell" (Russell, 2018), "Martinez" (Martinez-Sosa, 2022), "DeJonge" (De Jonge, 2014), and "Naafs" (Naafs, 2017) for temperature; "Russell" (Russell, 2018), and "DeJonge" (De Jonge, 2014) for pH.}

\item{data}{Data is presented as either peak areas ("raw") or a vector with the corresponding index values for the calibration ("index").}

\item{complete}{Logic variable. Indicates whether all brGDGTs in the original formulation are present in the data frame (TRUE, default value) or if some are missing. If TRUE the function will not proceed if any variable is missing, it will produce an error. If FALSE, the function will remove the missing variables from the equation.}

\item{na.ignore}{Logic variable. If FALSE (default) the function will turn any NA values into 0 for the calculations.}
}
\value{
The function returns the MBT'5Me value for each sample provided.
}
\description{
Applies published forward models based on linear functions for reconstructing environmental parameters using GDGTs
}
\references{
Russell, J. M., Hopmans, E. C., Loomis, S. E., Liang, J., & Damsté, J. S. S. (2018). Distributions of 5-and 6-methyl branched glycerol dialkyl glycerol tetraethers (brGDGTs) in East African lake sediment: Effects of temperature, pH, and new lacustrine paleotemperature calibrations. Organic Geochemistry, 117, 56-69.

Martínez-Sosa, P., Tierney, J. E., Stefanescu, I. C., Crampton-Flood, E. D., Shuman, B. N., & Routson, C. (2021). A global Bayesian temperature calibration for lacustrine brGDGTs. Geochimica et Cosmochimica Acta, 305, 87-105.

De Jonge, C., Hopmans, E. C., Zell, C. I., Kim, J. H., Schouten, S., & Damsté, J. S. S. (2014). Occurrence and abundance of 6-methyl branched glycerol dialkyl glycerol tetraethers in soils: Implications for palaeoclimate reconstruction. Geochimica et Cosmochimica Acta, 141, 97-112.

Naafs, B. D. A., Inglis, G. N., Zheng, Y., Amesbury, M. J., Biester, H., Bindler, R., ... & Pancost, R. D. (2017). Introducing global peat-specific temperature and pH calibrations based on brGDGT bacterial lipids. Geochimica et Cosmochimica Acta, 208, 285-301.
}
