% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pGDGT0.R
\name{pGDGT0}
\alias{pGDGT0}
\title{Percentage of GDGT0}
\usage{
pGDGT0(gdgt, complete = TRUE, out = "all", na.ignore = FALSE)
}
\arguments{
\item{gdgt}{Data frame with the peak areas or fractional abundances of isoGDGTs. Each row should correspond to a sample and each column to each isoGDGT with the appropriated name format.}

\item{complete}{Logic variable. Indicates whether all isoGDGTs in the original formulation are present in the data frame (TRUE, default value) or if some are missing. If TRUE the function will not proceed if any variable is missing, it will produce an error. If FALSE, the function will remove the missing variables from the equation.}

\item{na.ignore}{Logic variable. If FALSE (default) the function will turn any NA values into 0 for the calculations.}
}
\value{
The function returns the MI value for each sample provided.
}
\description{
Calculates the percentage of GDGT0 to Crenarchaeol for isoGDGTs presented as either peak areas or fractional abundances.
}
