% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RinTet.R
\name{RinTet}
\alias{RinTet}
\title{Degree of cyclisation of tetramethylated brGDGTs}
\usage{
RinTet(gdgt, complete = TRUE, out = "all", na.ignore = FALSE)
}
\arguments{
\item{gdgt}{Data frame with the peak areas or fractional abundances of brGDGTs. Each row should correspond to a sample and each column to each brGDGT with the appropriated name format.}

\item{complete}{Logic variable. Indicates whether all brGDGTs in the original formulation are present in the data frame (TRUE, default value) or if some are missing. If TRUE the function will not proceed if any variable is missing, it will produce an error. If FALSE, the function will remove the missing variables from the equation.}

\item{na.ignore}{Logic variable. If FALSE (default) the function will turn any NA values into 0 for the calculations.}
}
\value{
The function returns the #ringstetra value for each sample provided.
}
\description{
Calculates the #ringstetra index (Sinninghe Damsté, 2016) for brGDGTs presented as either peak areas or fractional abundances.
}
\references{
Damsté, J. S. S. (2016). Spatial heterogeneity of sources of branched tetraethers in shelf systems: The geochemistry of tetraethers in the Berau River delta (Kalimantan, Indonesia). Geochimica et Cosmochimica Acta, 186, 13-31.
}
